Added to the Body Demo model
Three spaces connected by an elevator shaft (currently missing an elevator. Won't add it until I have pushable buttons). Bottom floor is meant to be a sandbox area for messing around, first floor is for an introduction to the body, then the hand interactions, and then to how they combine in a parkour course. No clue what I'm gonna do with the top floor, but I'll find a use for it.